Changing of the regionMove function
Replacing the openNationOrRegion function with a slimmed down check to see if the latest happening has 3 links. If so, it moves to the target region of the latest move. If not, it ends. 

Also replaced the default refresh value to be pasted on the dossier report in function pageRefresh from 24 hours to 2 minutes-ish.